pyproject.toml
[tool.poetry]
name = "leyley"
version = "0.1.0"
description = ""
authors = ["enidisepic <143949913+enidisepic@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.3.2"
jinja2 = "^3.1.2"
loguru = "^0.7.2"
html2image = "^2.0.4.3"
pillow = "^10.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
